refactor(core/calendar): make calendar preset methods as independent module and fix some naming err
