+++ packages/core/src/calendar/index.ts
@@ -1,8 +1,2 @@
-export {
-  CalendarMethods as CalendarMethodsMoment,
-} from './presets/moment';
-export {
-  CalendarMethods as CalendarMethodsDayjs,
-} from './presets/dayjs';
-export * from './presets/typings';
+export * from './typings';
 export * from './calendar';
